requirements.txt
numpy==1.19.5
opencv_python==4.5.5.64
Pillow==6.2.1
